Fix GPDB/GPOS crash in debug builds
In Debug builds of GPDB, we try and shutdown the MDCache manually from
TerminateGPOPT. Accessing the SyncHashTable of the MDCache rquires
acquiring a spln lock, and to acquire that spin lock, we need a
PwrkSelf() worker object to make sure we don't deadlock etc. However at
this point all workers (created during OptimizerQuery) have been cleaned
up, and so IWorker::PwrkrSelf() returns NULL. This check is mute if
there is no worker, so we should guard against that case.